NOTE: This commit is a change in the backend interface! All implementations of the NDRegisterAccessor must be updated! Backends deriving from the NumericAddressedBackend do NOT need updating.
Fixed a bug in readAsync(), which could lead to calling a virtual function in the destructor of a base class (which is illegal). To circumvent this problem, the function NDRegisterAccessor::shutdown() must be called in all implementations of the NDRegisterAccessor in their destructor. See the description of shutdown() for more details (there are more details, so please do read it!).