Make dummy backend thread safe.
The dummy backend did not use any synchronization mechanism when reading
and writing. This meant that devices backed by the device backend could
show strange behavior when accessing them from multiple threads.

This patch introduces a mutex that is acquired for all read and write
operations. The mutex is not acquired when attaching callbacks or
setting address ranges as read only, as those functionality is specific
to the dummy backend and thus not exposed by the Device class.