** REMOVE FEATURE ** in LogicalNameMapping backend:
References in xlmap files to other registers are now restricted to constants defined in the same xlmap file. It is therefore no longer possible e.g. to dynamically deterime the channel number of a channel accessor from a variable.

Reason for the removal:

The feature is to my knowledge still unused despite it was implemented more than two years ago. The feature is notoriously complicated to maintain and somewhat stands in the way of the implementation of more important features.

If you are currently using this feature, please let us know quickly.