** REMOVE FEATURE ** in LogicalNameMapping backend (and in theory in other backend):
Removed register plugins, which could add information to the register catalogue and/or modify the data.

Reason for the removal:

The feature is to my knowledge still unused despite it was implemented more than two years ago. It overcomplicates the code and partially violates the abstraction in DeviceAccess.

If you are currently using this feature, please let us know quickly.