Avoid name collision of "label" types (issue #19).
There is a type "label" in the boost::unit_test_framework namespace, but
there also is a type with the same name in the sys/ucred.h system header
file that is in the global namespace.

This has the consequence that declaring
"using namespace boost::unit_test_framework" in the global namespace
leads to a collision.

This commit fixes this issue by moving the using declaration (and the
code depending on it) into an anonymous namespace so that there is no
name clash.