Force backend accessors to implement TransferElement::getVersionNumber(). The default implementation was dangerously wrong and could lead to strange effects (e.g. in ApplicationCore).
For a description how to do this implementation, see the Doxygen documentation of the function. For an example, have a look either at the NumericAddressedLowLevelTransferElement
(including its doWriteTransfer() and doPostRead()) for accessors directly reading/writing from/to hardware, or at the NumericAddressedBackendRegisterAccessor for accessors which use
other accessors for the actual data transfer.