Enable std::auto_ptr when compiling in C++ 17 mode.
std::auto_ptr has officially been removed in C++ 17. This means that
libc++ does not provide this template any longer, when compiling in
C++ 17 mode. By defining the _LIBCPP_ENABLE_CXX17_REMOVED_AUTO_PTR
macro, when can reenable it, even when compiling in C++ 17 mode.

While this is only needed on platforms that use libc++, defining this
macro should not hurt us on other platforms as it is very specific.