Change ReadAnyGroup::waitAny() to allow deferred processing.
The waitAny method now returns a notification object instead of a
TransferElementID. This way, the actual reading of the data update can
be deferred until later.

As a side effect, this makes the code cleaner, because now it is not
necessary any longer to call ReadAnyGroup::postRead(). Instead, the
notification object has an accept() method that takes care of this. As
this method is also called from the destructor of the notification
object, it is no longer possible to "forget" to call postRead().