Fixed open/close for SharedDummyBackend
Open and close were both throwing if the backend was already opened/closed. This is wrong and only the Device prevented the exception until now. But this does
not match with the new requirement that open try to re-connect, so the device cannot filter the call any more. All backends have to be fixed. This is the first one.

Test have been added.