wip #129: increment/decrement testable mode counter before/after device (re)opening
This prevents stepApplication() to return too early (e.g. right after start) when the device is not yet initialised.

Note: is has been verified that there is no race condition at the beginning despite the counter is only incremented during the test (by introducing an artificial delay at the beginning of handleException()). It works, because testable mode waits until all devices are opened. In testable mode, this check is now equivalent to the devices being also initialised (when the testableModeLock is obtained and the counter is checked).