Refactor Application::makeTypedConnection and FanOuts
Rearranged processing of fanouts and their consuming nodes. A list of
consuming nodes and implementations is passed to the fanout constructors,
so the nodes don't have to added later. This resolves conflicts in the
order in which accessors are decorated and added to the fanouts.

FanOut constructors could still use refactoring.