do not catch and rethrow numeric_cast exceptions
For one this makes tracking these exceptions virtually impossible, as they have a fixed what() string and hence no information about the variable which threw the exception can be added.

In addition, the concept of these exceptions is bad. The spec is already stating in C.4 that it must not be thrown in write operations, and it shall not be thrown by new implementations. Since these exceptions create problems, we need to move away from them as quickly as possible.