SubdeviceBackend 2regs/3regs mode: add support for data areas.
This fixes #177 (on github)

Note: This changes existing behaviour in one aspect: previously, the
address written to the address register was incremented in steps of 4
for subsequent low-level transfers of the same high-level register (like
a byte address). This does not make much sense, and likely no use case
relied on this behaviour. It should be noted though that this behaviour
was tested. The tests have been changed accordingly (and extended with
the new functionality).