fix LNM math plugin push parameters in device recovery (cf. #9961)
Previously, the MathPlugin would ignore incoming updates from parameters
if they were pushed in activateAsyncRead(), even if they were written
between the call to open() and activateAsyncRead(). This is a severe
issue, since ApplicationCore is first recovering all device registers
and then calling activateAsyncRead().

Note: The fix needs also a change in ApplicationCore, since it
previously writes the recovery values with the VersionNumber from the
last write, which was before the call to open().

Another solution to this problem would be to accept variable parameters
in the MathPlugin which have been written at least once after creating
the backend (in contrast to after the last call to open()). This
solution was rejected, since it would cause the MathPlugin's target
register to be written many times during device recovery (once per
parameter plus one time for the main value).