Subdevice: wait for handshake after write
Change the Subdevice backend in 3regs and areaHandshake such that it
waits for the handshake signal that the transaction is complete before
returning the write operation.

The previous behaivour was to wait before commencing a new transaction
until any previous transaction is complete. This created issues if the
remote state machine for the handshake system (inside the FPGA) is reset
through a register write, since there is no way in software to determine
whether the reset may interrupt an ongoing transaction.

With the new behaviour, it is perfectly safe to write the reset register
right before or right after the subdevice write.

See redmine #9959 for further details.