fix wrong type information for unsigned int
The number of digits reported for signed and unsigned integers was
partially equal, which is wrong due to the additional sign character.

This introduces a slight behaviour change, which likely will not be
observed in real applications. ApplicationCore bases its type guessing
on this information but already uses the correct thresholds.