Use correct data type when calling NumericAddressedBackend::read.
There are two versions of NumericAddressedBackend::read, and neither of
them expects values of type size_t for the first two arguments, so
calling it with such values results in an ambiguity which version should
be used.

By making sure that we call it with arguments of type uint64_t, we can
resolve this ambiguity.