hotfix: rethrow old exceptions by math plugin after recovery
Note: This change is likely to be replaced with a better implementation
(with potential side-effects, so not to be released as a patch).

The problem solved with this commit is that the math plugin with push
parameters prevented the recovery from a runtime_error, because the old
runtime_error was still present in the parameter queues after re-opening
the backend and was hence re-thrown in the next write operation, placing
the backend again into the exception state. This is now solved by
clearing the queues in the open_hook.

A better implementation might be to not use wait_for_new_data for the
parameters (even if push parameters are enabled), so there are no queues
at all which could store the exception. wait_for_new_data is in
principle not necessary, since the math plugin never executes blocking
reads on the paramters. Instead the push parameter variables will
trigger the update of the math plugin in its postWrite directly. As a
side effect, the math plugin push parameters would work independently
from activateAsyncRead(), which probably can be considered cleaner and
more intuitive.