fix(xdma): backend must not close itself on error
The XDMA backend called internally close() when a recovery of an
exception was attempted. In particular if the recovery was still
unsuccessful, the device was then left in the closed state leading to
logic_error exceptions in subsequent read/write operations. This caused
ApplicationCore applications to terminate immediately after an exception
of an XDMA device was seen.

This is now fixed by using a dedicated flag for the backend opened
status which is independent of the underlying interface status.