fix: don't throw logic_error in TransferGroup directly (#14286)
This is no longer necessary, since the subsequent handling of all
relevant exception types is appropriate also for logic_errors. It was
also an issue, since in ApplicationCore the ExceptionHandlingDecorator
prevents initiating transfers if the Device is temporarily closed
while executing an init handler, but the TransferGroup throws the
logic_error anyway.