fix: LNMBackendBitAccessor replaces its target with copy decorator
In replaceTransferElement, the LNMBackendBitAccessor was replacing its target with a CopyDecorator to the target,
also for a single instance. As the CopyDecorator skips the call to preRead() of the target, this ended up in the
preRead() never being called. As a consequence, the inTransfer flag was not on and the rest of the read operation
was skipped.